Time: 0 ms (100%), Space: 8.2 MB (76.45%) - LeetHub

+++ 1797-goal-parser-interpretation/1797-goal-parser-interpretation.cpp
@@ -1,11 +1,11 @@
 class Solution {
 public:
     string interpret(string command) {
-        int commandLen = command.length();
+        size_t commandLen = command.length();
         string answer;
         answer.reserve(commandLen);
 
-        for (int i = 0; i < commandLen; ++i)
+        for (size_t i = 0; i < commandLen; ++i)
         {
             char ch = command[i];
